backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pinecone==4.1.0
google-generativeai==0.8.1
pydantic==2.5.0
python-multipart==0.0.6
rank_bm25
mangum==0.17.0